Broke out msgf into its own biocontainer
