+++ environment.yml
@@ -4,7 +4,6 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-  - bioconda::msgf_plus=2020.03.14
   - bioconda::hardklor=2.3.2
   - bioconda::kronik=2.20
   - bioconda::dinosaur=1.2.0
